requirements.txt
selenium==4.17.0
pytest==8.2.0
pytest-order==1.1.0
allure-pytest==2.14.0
webdriver-manager==3.8.6
